requirements.txt
fastapi
uvicorn
transformers
torch
psutil
python-multipart
Pillow
openai-whisper
pyttsx3